refactor : 인터페이스 분리 및 컨트롤러 V2로 이관
