+++ src/main/java/app/bottlenote/global/security/jwt/TokenValidator.java
@@ -1,0 +1,10 @@
+package app.bottlenote.global.security.jwt;
+
+import io.jsonwebtoken.Claims;
+
+public interface TokenValidator {
+
+    Claims validateAndGetClaims(String idToken, String expectedNonce);
+
+    String getEmail(Claims claims);
+}
